Updated the way rooms are created
Rather than using coordinates, I am taking a different approach where I use the number of rows and columns as well as the origin position of the room in order to handle splits. This is to keep the grid nature of the generation. Currently there is an infinite loop while checking which rooms to split